requirements.txt
torch[cuda]
torchvision[cuda]
transformers
numpy
Pillow